Move towards one package list for a defined distribution
The previous version of the instructions suggested an xargs
invocation that didn't work: a whitespace-separated '|'
in conf/packages caused every package to be selected.

It also seems like a good idea to have package lists and
instructions that are specific to a known-to-work
distribution, and add others if they are contributed.