Make install_perl_modules exit in error on failure
Previously, install_perl_modules would exit with 0 regardless
of whether installation of the modules succeeded or failed;
this meant, in particularly, that when called from the
install script with 'set -e', it's possible to miss that
the module installation failed.  By adding 'set -e' to
install_perl_modules, we should get a useful exit code.