A fix for the (bizarre) text entry problems on Android
This is a really very strange problem.  When testing the
application on an Android 2.3.5 device, there were two
problems with the <input type="text"> box on the front page:

 * The left and right arrow keys wouldn't move within the
   text of the box - instead they would switch the focus
   highlight to another element.

 * Holding down the letter key to get a number (e.g. 'w' to
   get a '2') would end up with 'w2' being entered rather
   than just '2'.

I couldn't reproduce this in a newly created project, so
by introducing more and more of the code from the FixMyStreet
phonegap app into that project isolated the problem to
the CSS for an unrelated element.  In short, the text input
boxes will work correctly after removing 'position: fixed'
from the mobile-info class.  That class is only ever used in:

  <div class="mobile-info">
    <a href="about.html">i</a>
  </div>

Why this could possibly cause the problem above, I can't
guess.  However, removing it makes the Android application
usable again, and we are shortly going to redesign the
application in any case, so the sub-optimal position of the
'i' button is something we can live with.