Set up internationalization with Transifex
gettext-extract needs to be changed to instantiate values
in the Plural-Forms header line, otherwise on pushing the
source pot file (locale/FixMyStreet.po) it fails one of
the checks that Transifex runs (msgfmt -c).

This commit also adds the .tx/config configuration file,
which points to the right project on Transifix and
specifies where the source and translated files are in
the project structure.