Remove forked fonts and base CSS from Stevenage.
Revert Stevenage's base stylesheet to inherit from FixMyStreet's,
removing any unnecessary forked files, but leaving in the font overrides
and change of front page button colour.

Fix the $primary_text colour to be visible over $primary (this then
removes the need for the colour override of the base part of 1162d5c8).