A script to bring a FixMyStreet database up to date.
It will examine the current database to see at what state it is,
and optionally run the newer schema SQL files in order to bring
the database up to date.

At the 'adding bodies' stage, it will also query MapIt to fill the
name column of the new body table.