added diagram showing how geocoder/mapit/database
yet to write the text supporting this diagram, but I understand this is a
fairly heavyweight diagram to back up the general description of how
the geocoder, mapit, bodies, and contacts fit together. Depending on the
target reader, we should or should not be showing the separate database
tables, since this is not quite how they are presented in the admin
(of course). Comments welcome, obviously.