Do fewer things for a dev installation.
If the development parameter is passed to the install-script, don't
install a crontab, don't set up nginx, don't force the repository
to a particular version, and don't set up an application server.

Also only create a general.yml file if one doesn't already exist,
and add some locales that the test suite assumes are present. And
run the db schema update script.