[Bromley] Fix right click links in IE8
When IE8 encounters a link with block level elements inside it (i.e.
elements with `display: block`) then it won't give the option to "Open
in a new tab" when you right click the link.

This fix simply makes all elements inside the link inline, which fixes
the problem.

This is a Bromley-specific fix.

Closes FixMyStreet-Commercial/issues#100