[Zurich] Remove phone number requirement for mobile app reports
The phone number was made mandatory in #541. To make this change
compatible with older versions of the mobile app I've had to drop
server side validation of the phone number for requests that come in
from mobile apps.

To determine if it's a mobile request I'm checking the 'service'
parameter in the request, which is only set in requests from the mobile
apps [1].

The phone validation will still happen client side for the newer
versions of the app, and the web app will continue to work with the
server side validations.

[1] https://github.com/mysociety/zurich_mobile/blob/1fec2bc21dad9326819105cdbf5d47183e7503ba/www/js/models.js#L75