Work around Chrome bug with bottom fixed navbar.
The position:fixed bottom navigation does not stay fixed in Chrome, it
appears to scroll with the content (though the link targets remain and
swapping tabs fixes the display, so I'm pretty sure it's a bug). The bug
is logged in Chrome's tracker at:
https://code.google.com/p/chromium/issues/detail?id=311513

To work around it, the z-index has been removed from .container; this
would mean that the sidebar notes all disappeared under the map, so they
have been jigged to appear on top of the main content div instead.