Don't display categories for deleted bodies
When a body has been marked as deleted (currently only exposed in the
Zurich cobrand's admin) the categories for that body should not be
available on the report_new form.

Prior to this change there was an error running the
'zurich-overdue-alert' script, because it wasn't expecting the
'bodies_str' field to have any entries in the format 'x|y,z', this
prevents this problem from happening in the future (though it doesn't
fix any existing entries with this issue).