[Zurich] log details for closed_overdue
see mysociety/FixMyStreet-Commercial#459 the new requirement is for
"Innerhalb von fübeitstagen abgeschlossen" - a report is dealt with
(abgeschlossen) as soon as the public state is Beantwortet (db: fixed -
council or closed) or Unsichtbar (en: Invisible, db: hidden).

Have added this as {closed_overdue} but have retained the logic for
{subdiv_overdue} logged in the %extra hash for now, in case it is
needed.

Note: as per Matthew "as with confirmed/in progress this should be 6 -
it's five days from moderation so is most simply just done as being six
since creation"