[Zurich] admin template to respect closed_overdue
This value is now cached, but the template calls the function,
calculating it anew, so for old reports (closed within SLA) the
problem_row.html template showed these as overdue.  Instead, we point
the template at a function which first checks the cached value.

(Attempting to roll this functionality into the .overdue() method itself
caused a fair amount of collateral test breakage, and it seems safest to
extract the caching function for now.)

Closes https://github.com/mysociety/FixMyStreet-Commercial/issues/462