Re-enable dashboard tests
The tests have been commented out due to long-running failure.  It looks
like the failures happen if working copy MAPIT_URL is not pointing at
live site. Have updated with an override_config setting this to live
site (which is obviously suboptimal, but seems to be widespread practice
across t/ at the moment.)