enable generation of cobrand specific po files
This is to stop the main FixMyStreet .po file being filled up with
strings that only apply to a single Cobrand that are not relevant to
people running their own fairly vanilla version.

Restrict the list of templates that we pull strings out of by generating
a list of all the ones in currently translated cobrands for minimal
impact. Use that to create the FixMyStreet.po file in gettext-extract.

I've left the existing cobrands in but at some point we should split
these out.

Add a specific script for generating Cobrand .po files which extracts
strings from the Cobrand's templates. This is then translated and the
translated file place in the locale/$lang/LC_MESSAGES.

The the merge_cobrand_po script is run which will create an
autoCobrand.po file which has the Cobrand strings and the FixMyStreet
strings. This is then the domain used by the Cobrand.

Any cobrand specific strings from the Perl code are still in the
main FixMyStreet .po file but there should be relatively few of them so
it's less of an issue.