Allow comments to be confirmed with no confirmation
Add never_confirm_updates setting to cobrand that, when set to 0 means
that comments are confirmed with no confirmation step. Essentially
behaves as it would if the user was logged in.
Also add a flag to the flash that indicates the comment was accepted to
you can display a message on the report page if required