Add a searchable dropdown of clinic names to postcode form
uses select2 for the dropdown and a cobrand call to fetch the list of
report titles as the clinic names.
Relies on a redirect in httpd.conf to make the non JS version work.

Still leaves the postcode search in place as select2 won't accept custom
terms if you base it on a select box, which we want to do so it works if
JavaScript is disabled

Fixes #723