Fire a google analytics event on report completion
At the moment it's hard to use google analytics to track when someone
has completed a report due to the various redirects which can occur too
quickly for the analytics javascript to fire. So now we fire an event if
we've set a flash value. This also allows us to track how they created
the report.