Add a who are you trying to contact option to contact form
In order to try and cut down on people mistakingly using this form to
try and contact their council all some 'who are you trying to contact'
options to the form that display appropriate messaged unless the 'FMS
team option is selected.

Add the validation code for this in the cobrand module.

Fixes #41