alter pave url sent to analytics on report completion
In order to allow us to do funnel analysis in google analytics we need
to send a differnt page url when a user completes a report. To do this
we add a report_created=1 to the query string on either the report page
or the report confirmation page for logged in and confirm by email users
respectively.

There's a bit of complication as we don't want to set a session cookie
everywhere as that will break caching so we set a stash variable,
although if the user is logged in this is controlled by a variable in
the flash as there is already a session cookie at this point.

also, removes the code that sends an event upon report completion as
this was no use for funnel analysis.