Fix DB warning being shown by zurich-overdue-alert
A raw DateTime object was being passed as a parameter in an SQL
query, which was causing undefined behaviour and a warning
every time the Zurich overdue problems script was run.

The DateTime is now converted into a format suitable for use in
the query and the warning has gone.