Parameterise query for load_problem_or_display_error
By default `load_problem_or_display_error` only allows reports
that are owned by the current cobrand to be displayed.
Allowing the base query to be passed in as a $problems parameter
means that the function can be reused in places where the base
set of problems isn't limited to those owned by the current cobrand.