Give problem confirmation page its own URL
Previously report confirmation page was only shown in
particular circumstances (generally if the user wasn't logged in)
and at a single-use URL (the token link from the confirmation email).
It will be advantageous to show the confirm page at other points so
this commit exposes a new URL /report/confirm/<id> that shows a
report's confirmation message.