Move bromley from report hiding to moderation
The feature to hide reports is still enabled, as it is used by other
`users_can_hide` bodies.  As we want to roll out moderation to Bromley,
in this commit we hide the functionality *only* for Bromley.  Of course
if we migrate those other users to this moderation method, then we will
need to make equivalent change to general template, and remove the
functionality from Report.pm.