Better spotting of signing in on /auth form.
If your browser autocompleted form fields, you could fill in the signing
in part of the form but still be sent a confirmation email.  This commit
will now default to trying to sign in if the sign in button is clicked
or there is data in the signing in password field.

Fixes #816.