Bromley. signature and FAQ tweaks
As per mysociety/FixMyStreet-Commercial#524 to stem the flow of
people responding to verification emails by clicking the support
email at the bottom of Bromley's signature, we are instead redirecting
to the FAQ -- which is also updated to contain the support email
address, so the people can still find a contact if they need to.

Also tweaking moderation FAQ line, in preparation for moderation
trial.