[Bromley] disambiguation for White Horse Hill
This street is on boundary of Bromley with Greenwich, so a simple
disambiguation of $town = "chislehurst" falls foul of UKCouncils->
area_check.

It may be worth overriding area_check in general, but in mean time,
setting a specific postcode (on the Bromley side of the administrative
boundary) allows the map to be seen.