Replace from_council with from_body in SeeSomething cobrand
'from_council' was renamed to 'from_body' and turned into a relationship
in b2aa34b Feb 2013, and the seesomething code was missed.

Though the FMS::DB::Result::User object would (correctly) throw an error
if an invalid method is called, *Catalyst* idea of a user object wraps this
in a Catalyst::Authentication::Store::DBIx::Class which uses AUTOLOAD to
delegate methods and swallows the error.  e.g. this code always returns
false, hence the code path is never run.

(I think this is a bug and have opened:
    https://rt.cpan.org/Ticket/Display.html?id=97640 )

Added test case and (hopefully) corrected template, will need a check
against site running copy with recent data to check that nothing looks
squiffy.  We also had to change the %council_details stash, which was
based on the old code, not sure if suggested fix matches original
intent, but looks plausible.  (As discussed with @dracos @stevenday
this is slightly repeated, and should arguably use ->{bodies} but that's
a possible future FIXME)