[East Sussex] pothole field input and display
This is a temporary request, which we hope will be superseded by full
integration in due course.

A command is documented in the cobrand to set the "extra" on the
contact info for the Potholes category.  This is then displayed by
the UI.  The default process_extras places it in extra, for potential
future processing.

As this is a temporary fix, we are calling this field "detail_size", as
this is automatically included in "Details" and therefore also sent in
the email.

(This means that we may need to update data later on to call this
"pothole_size" or similar in future?)