[Collideoscope] re-branding
 - new reporting form questions and markup
 - define collideoscope brand colours
 - document where colours are used, Sass colour rework
 - remove custom colour declarations from smidsy base.scss and layout.scss
 - remove squirrel logo
 - remove mention of squirrel logo
 - remove mysoc tab from smidsy, reword nav links
 - sort out display of smidsy nav bar and logo
 - smidsy body background images
 - smidsy footer
 - cabin typeface for smidsy
 - quick layout fixes for smidsy reporting page
 - sidebar for smidsy .twothirdspages
 - smidsy homepage, without any images yet
 - Add extra options for customising #postcodeForm geolocation link

This commit forks fixmystreet’s header.html and the base fixmystreet.js,
to add hooks for customising the look of the geolocation link.

It also removes the space that’s inserted before the loading spinner,
and replaces it with margin-left in _base.scss, which avoids the visual bug
of a link underline extending past the end of the geolocation message.

Both these changes should probably be merged into master at some point.

A further improvement might be possible through the combination of the
`fixmystreet` varaible introduced here, and the one defined on map pages
in templates/web/base/maps/openlayers.html.

 - marker icon for geolocation link on homepage
 - smidsy homepage polish, including how-it-works images
 - smidsy replacement for "click map" banner
 - smidsy list of incidents, easter egg for empty list
 - smidsy incident list
 - improve contrast of homepage search box focus state
 - add sponsors page, link to it from other pages
 - smidsy sponsor page, non-functioning form
 - amend smidsy statistics page content