Add extra options for customising #postcodeForm geolocation link
This commit forks fixmystreet’s header.html and the base fixmystreet.js,
to add hooks for customising the look of the geolocation link.

It also removes the space that’s inserted before the loading spinner,
and replaces it with margin-left in _base.scss, which avoids the visual bug
of a link underline extending past the end of the geolocation message.

Both these changes should probably be merged into master at some point.

A further improvement might be possible through the combination of the
`fixmystreet` varaible introduced here, and the one defined on map pages
in templates/web/base/maps/openlayers.html.