Smidsy. fix injury box show/hide logic
In safari, you can't get .value for a radio group, so we use the
jQuery :checked selector instead.

Also, it looks like .slideUp isn't honoured if the element is
already hidden (as per FMS logic on /around? page).  If we just
.slideUp and .hide then the box would simply get hidden, but we
can add the hiding as a callback, and this seems to do the right
thing in all cases.