Pass cobrand into send_email_cron when sending SendReport emails
Without this, email sending within a cobrand will get the right main
email template (submit.txt) for the email - because it uses the
cobrand inside get_template - but it doesn't use the cobrand for
things like the email signature, because you have to pass it into
send_email_cron manually to make it cobrand-aware.

This probably hasn't been noticed before because most cobrands are
called Something FixMyStreet, but Collideoscope is different and hence
it caused a problem there.

Closes mysociety/FixMyStreet-Commercial#621.