Smidsy. fix updates bug caused by bad typing assumptions
Smidsy followed EHA practice of a hashref for extras, instead of
an arrayref.

That's convenient, but it turns out that EHA doesn't expose updates,
so there were no hooks for this.  As we don't actually have extras
for updates, am just returning in this case.

see https://github.com/mysociety/FixMyStreet-Commercial/issues/632