EastSussex. z-index fixes for IE7
The #map_box div wasn't clickable, so we had to reset its z-index from the
broken -1 to 0.  IE7 doesn't like setting z-index to 0 in javascript (as it's a
falsy value) so instead we set the z-index of parent and sibling elements (body
and site-header) to 1 to make sure they are visible.

position:relative in base.scss is to make sure that the header is now
shown (similar to @6296cb33)

See https://github.com/mysociety/FixMyStreet-Commercial/issues/639