Fix js-disabled submission in Chrome/Firefox
Fixes #932, which is caused by HTML5 validation firing for hidden
elements.

Sets #mapForm to novalidate by default, and reverts this in
fixmystreet.js.

(Note that the attributes 'validate' and 'novalidate', which are
honoured by the browsers' HTML5 validation) aren't to be confused with
the class (class="validate") which is presumably only picked up by
javascript, and therefore not problematic for the no-js case.