Fix JS-disabled submission in Chrome/Firefox.
Fixes #932, which is caused by HTML5 validation firing for hidden
elements, by setting #mapForm to novalidate by default, and reverting
this in fixmystreet.js.

Note that the 'novalidate' attribute, which is honoured by the browsers'
HTML5 validation, isn't to be confused with the class (class="validate")
which is picked up by fixmystreet.js, and therefore not problematic for
the no-js case.