TO REBASE flatten label and add tooltip
See https://github.com/mysociety/FixMyStreet-Commercial/issues/642

Try to avoid label looking like a button by flattening it
(e.g. removing box shadow).

The cursor icon change is helpfully set by the fact we're in a <label>.
As that might be a feature, instead of disabling the change, I've added
a <title> attribute to at least generate a tooltip.