Enforce per-category photo requirement on new reports
A new step has been added to the photo upload process that
ensures a photo is present if any of the categories chosen for
the report require it.

If the photo is missing an error is displayed to the user in the
same manner as if the photo upload was invalid.