Zurich: Don't treat Problem->extra as an array
The 'extra' field on Contact and Problem is used quite differently
across various cobrands, in an incompatible way. Some cobrands treat
the field as an array of hashes, whereas others treat it as a single hash.
This commit will, when the Zurich cobrand is active, skip over the logic
that sets Problem->extra as a single-element array when Contact->extra is
present to avoid a crash when confirming a report.