Add Extra role to ease use of {extra} field.
Historically, the extra field has been used in two different ways by
different cobrands, both as a list (e.g. Open311 category fields) and a
hash (e.g. the Zurich cobrand).

This commit consolidates usage, adding an API to make use of the field
easier and always returning a hash for the code to use. Fixes #1018.