[Harrogate] show warning banner for NYCC categories
Harrogate asked for NYCC managed issues to be linked to that
council's website.  We don't prevent the issue from being logged,
but do show a warning banner.

As this is at Harrogate's request, we only do this for the Harrogate
cobrand, not the whole site, using a new cobrand hook
`munge_category_list`

Unfortunately, categories are currently passed to the template as
plain text, e.g. not structured data. As a simple implementation,
we append "( NYCC)" to the appropriate category names.

JS could then strip the names back in future, if required (though
we're not doing this.)  Instead we add a jQuery hook to show the
most appropriate link provided.

See https://github.com/mysociety/FixMyStreet-Commercial/issues/703