[Zurich] additional closure statuses work
Requirements:
=============

1: DM to be able to mark a problem as Wish, Extern, Not contactable,
Hidden, Jurisdiction unknown

2: SDM to be able to mark a problem as Not contactable.

3: DM to be able to add information to such a problem, and trigger
emails to external agencies, and to user, closing the problem at the
same time.

4: Closed reports must be categorized as Wish/Extern/NC/Hidden/JU for
statistical reporting

5: Non-closed reports must appear as open, to allow DM to see them in
the list, similar to what is currently done for Ruckmeldung Ausstehend
(Feedback outstanding).

5b: however "Hidden" problems must be immediately hidden from public
view.

Proposal:
=========

a) Requirement 1 will be carried out in exactly the same way by
DM.  e.g. Category dropdown select

b) Requirement 2 will be carried out in exactly the same way by SDM.
e.g. "Not contactable" button

c) BUT, for a), c), the category will be diverted to "Ruckmeldung
Ausstehend" (Feedback outstanding), and metadata will be saved with the
appropriate closure type.

d) NOTE: this means the report will still be visible to public as "In
bearbeitung" (Except for Hidden reports, as per 5b)

e) The extra fields will be presented on the "Ruckmeldung Ausstehend"
page, just as defined in the PDF, depending on the metadata.  e.g.
    - Wish: Zustaendige Stelle + Message
    - Extern: External body + Message2
    - Not contactable: (no extra info)
    - Hidden: (no extra info)
    - Jurisdiction unknown: (no extra info)

f) When the DM publishes the answer (e.g. "Antwort Veroeffentlichen"
button), the state will normally change to Fixed. BUT in the new cases,
the metadata saved in step c) will be checked, and the report will
change state to Wish/Extern/Not contactable/Hidden/Jurisdiction unknown,
as appropriate.  The report will now be visible/hidden as previously
discussed depending on the new state.

g) If the report is edited again (in /admin/report_edit) it will be in
the relevant state, and any extra fields will be visible as per e).
Those fields may be updated.
Alternatively, the state may be changed, possibly triggering a) again,
and clearing any metadata.

 - Fix javascript interactions 690
 - show new statuses in summary
 - fix photos on SDM page
 - fix Planned => Planned+metadata transition
 - stop "assignation .select" js affecting templates
 - fix redirection for Extern and generally

See mysociety/FixMyStreet-Commercial#672, mysociety/FixMyStreet-Commercial#690