[Oxfordshire] Refresh map as soon as filters are changed on empty /my
By default the map JS doesn't get included on /my if the map is empty, so
there's some slight duplication of the JS from the previous
commit to make the filters submit the form as soon as they're changed.

This should be included in the base template if/when the filters make their
way there.