[Oxfordshire] Move council info into own section of report page
Oxfordshire reports have a council Reference Number, and an
estimate of how long they'll take to be fixed. To make this
information easier to parse, we've broken it out into its
own section, between the problem description and updates.

This looks ok with the current Oxfordshire map page layout, but
will need updating when issues/commercial/687-oxfordshire-full-screen
is merged.