[Zurich] Tweaks to redirect
See mysociety/FixMyStreet-Commercial#690

I understand redirection to summary page was introduced at ZWN's request,
to make it easy to
process reports and then quickly move onto the next one.

However, during testing now, Tobias has mentioned this is a) slowing
things down and b) confusing because it's not obvious that the report
has been saved.

I've tried to address (b) by adding the "Aktualisiert!" message when
you are redirected.

Also, for (a) I've removed the redirection from a few cases, and
disabled it for superuser.  If need be, I can remove redirection from
more (or all) cases.

Fix Official answer/Reply to user

 - hide label for status update on state change
 - correct wording on button for closure (single wording)
 - correct wording on label for user reply (either "Official answer" or "Reply to user" as appropriate)
 - Make sure the official response texts are shown for edit/static as appropriate, and test.
 - javascript improvements
 - honour public status update for Extern/Wunsch too
 - don't show public message for Wunsch
 - Ignore all other fields when rotating photos.
   (See mysociety/FixMyStreet-Commercial#718)