Squash c#690 tweaks to redirect
I understand redirection to summary page was introduced in
7c1df92482fd5e69d by @dracos at ZWN's request, to make it easy to
process reports and then quickly move onto the next one.

However, during testing now, Tobias has mentioned this is a) slowing
things down and b) confusing because it's not obvious that the report
has been saved.

I've tried to address (b) by adding the "Aktualisiert!" message when
you are redirected.

Also, for (a) I've removed the redirection from a few cases, and
disabled it for superuser.  If need be, I can remove redirection from
more (or all) cases.