[Zurich] WIP noscript_map using percentage widths
The default noscript_map uses set pixel dimensions, which makes it
really hard to display at different sizes (eg: in print layout).

This commit alters the noscript_map markup and styling to use floated
tiles with percentage widths, and then percentage positioning for the
marker pins.

The marker appears to be positioned in the wrong place, but Hakim is
going to look into that. (Something to do with pin.px and pin.py)