[fixmystreet.com] Unresponsive bodies page/warning
Add a special category email address used to show an unresponsive
message and page when trying to report in that category. Add a "Refused"
send method for setting a whole body to be unresponsive and show the
message immediately, not on category selection.

Factor out category template, and put at top where needed.