Stop SSL verification warnings on commercial staging
This adds to commit 90fb1b7 (f9f0906 in commercial-staging) where
hostname verification was turned off on staging sites. Because of
this, we started getting warnings from IO::Socket::SSL that we
need to explicitly turn off SSL verification if we really want this.
As per the comment on 90fb1b7, this is perhaps a bit of a stopgap
rather than a full solution.