Base 64 decode uploaded images if required
The new Zurich app has to base64 encode some images into a multipart
request so that it can (ab)use Phonegap's FileTransfer plugin to send
more than one image in a single request to /report/new/mobile.

Catalyst's Request::Upload module doesn't look at the
Content-Transfer-Encoding header inside mulitpart requests to do
this automatically, so we perform the decoding manually when we
process the images in the PhotoSet model.