Drop show-admin-notes class and show notes everywhere
.show-admin-notes was a class that could be applied on a cobrand
basis to turn on or off the showing of admin notes and error
messages. This was introduced when the concept of admin notes
was added, so that existing cobrands could opt-in to the
functionality. Now, every cobrand uses it, so there's no need
to make it optional.

This commit removes the CSS styles based on it, so that errors
and notes always display; removes the addition of the class to
admin pages, and removes the conditional javascript that enabled
admin tooltips based on the class.

Closes #1161