Fix sidebar running over the footer on alerts page
This page used the .sticky-sidebar class that is used in the help
section to present the in-page navigation of a lot of text. There,
there's no danger of the sidebar being scrolled over the footer
because it's very short. On this page though, there is.

This commit removes that class, so that the recent photos are fixed
in position with the top of the content. I don't think this is a
great loss for usability or design. It also removes the alertindex
class, which was only used to work around this issue.

Closes #1168