[Collideoscope] Remove js for stats19 checkbox and replace with a link
We need to process this filtering functionality properly, not just
artificially attach it to the "show old" button already present.
To do that, it has to be a parameter that's passed to the controller,
and following the examples of the "hide pins" and "show old" controls,
the easiest way to do that is to by making it a link.
If we naively make this a form element, things get confused, because
there's already a hidden form wrapping the whole sidebar.