[Collideoscope] Revert to using on/around _map_list_items.html
Collideoscope originally used a new template:
combined_map_list_items.html that was referenced from
tabbed_lists.html on it's own, because it doesn't have the two tabs of
different lists of problems on or nearby the map that FMS does.

However, the original templates were used in other places, like when
rendering new lists of pins for ajax responses. This commit goes back
to the original templates, using only the first one:
on_map_list_items.html to show a combined list, so that the javascript
behaviour is preserved. around_map_list_items.html is included, but
left blank, so that it's clear that we don't want to render anything
for it. (Otherwise the ajax controller would waste time building HTML
from it and sending that to a frontend that wouldn't use it).