[Collideoscope] Make stats19 on/off link work via javascript
This required a new template override for the map on smidsy so that we
could include a tiny bit of javascript to pass on the stats19 setting.

Apart from that, it works by using FMS' existing pin refreshing
functionality, forcing a refresh on page load to get pins that respect
the setting (otherwise, FMS' default which does similar will override
the static HTML pins).