Don't set whole background style in issue-list css
The use of the background shorthand was causing cobrands which
override this to have unexpected styling issues because it
only set a colour, not the other background properties.
This commit makes it more specific to just use background-color.

Also fixes a case in Greenwich's css where this was being overriden
in a similar way to just use background-color as well.

Fixes #1238