Fix questionnaire image sizing in Firefox.
A 100% max-width img within an inline-block within a fixed width wrapper
behaves differently in Firefox to other browsers; other browsers use the
width of the fixed width wrapper, whereas Firefox uses the width set for
the img. Given an identically set-up div, all the browsers use the width
of the div rather than the fixed width wrapper, so Firefox appears to be
behaving more consistently.

The easiest solution is to change the inline-block to display: block, as
the inline-block is alone within a floated fixed width div. Fixes #1243.