Combine some base/fixmystreet templates.
The aim is to combine base and fixmystreet as much as possible,
providing enough template/styling hooks for any cobrand. This
commit removes the need for separate base/fixmystreet templates
for the front page, almost all of the /around templates, the
open graph headers, and the low level item list templates.