[Bromley/Greenwich/Oxfordshire] Tidy overrides.
Have the around, my, and reports pages use the same low level template
for individual list items, which can then be overridden for these
cobrands, meaning they no longer need to override higher level
templates.

The logic in 2b9178d was incorrect; fix this so that the correct
information is included in those templates.