training pages for FMS (three types of user)
these pages for MakeMyIslands team training
should be general enough to be useful to *any*
new team needing such introduction. The three
types of user are citizen, staff, and admin.

There's a problem with the "View as slideshow"
magic (the stacking context of the modal is
wrong so the close button doesn't work)