selectively replace FMS with MakeMyIsland in training/*
In the training for citizens and staff users, the use
of "FixMyStreet" for the name of the site is going to
confuse (but not for the admins, who know they are running
an instance of the FMS platform). So for now, I've changed
the words.

Currently the training pages are only for the MakeMyIsland
installation, so this seems like a practical way to start.

Screenshots are from MakeMyIsland.

Maybe, if there's demand, this can become either switchable
or custom versions generated as separate jekyll sites off
the vanilla FMS training one (such as in the previous commit).