Fix cobrand restriction of My/Nearby.
5c79337 simplified a bit too far, as after then a particular cobrand
could in Nearby and My only filter reports to a particular body, not
any other criteria. To fix this, introduce more generic functions in
the default cobrand to allow more flexibility.

Make sure a few tests delete their bodies fully so that new tests
pass when run as part of the suite.

Fixes #1289.