Show remembered uploaded photos when Back used.
The uploaded_files solution in cbdfcad6, to show already uploaded files
stored in upload_fileid, worked if you submitted the form to the server
and had it returned, but did not show remembered photos if you used the
browser's Back button to return to the form from another page.

Instead, always read the data from upload_fileid, which works whichever
way you have come to the page. Fixes #1332.