Skip sending if any body marks it for skipping.
If a report was sent to multiple bodies, one of which was set to skip,
(e.g. temporarily turned off due to receiving issues), then the code
would still try and send the report due to the other bodies, leading to
multiple sendings and an error for the skipped body.