Absolutely position map pages.
This simplifies the page layout and code. We add a new #map_sidebar for
sidebar positioning, and a $mappage-header-height (defaulting to 4em)
variable. We no longer need:
* position_map_box;
* map_fix;
* special visual centre calculation; or
* custom pan zoom positioning.

As the sidebar now scrolls by itself and nothing can extend outside:
* The big-green-banner is set to 100% width (with negative margins no
  longer required);
* For cobrands that have an overhanging sidebar (e.g. Zurich), we add
  a .with-notes variant that adds a second column to the sidebar,
  perfect for containing secondary notes about the form.

The resizing of the banner also required some tweaking to the size of
the chevron-white SVG images.

Update the $.fn.drawer() method for the new layout. Works well in IE7+8,
and degrades gracefully in IE6.

For IE7, we remove the existing z-index hacks, and put a new, simpler,
one in, and fix a CSS bug that was hiding the map.

Various cobrands are updated to adapt to the changes.