Updates to cpanfile and code for perl 5.20/5.22.
The following modules had bugs that have been fixed for working in
recent perls:
* List::MoreUtils
* Guard
* PadWalker
* aliased
* URI
* Convert::NLS_DATE_FORMAT

The CGI module was removed from core in 5.20, so include it in the
snapshot (I don't think it's actually used, but is a dependency).

"{" needs to be escaped in regular expressions, and ~~ should not be
used.

Fix some tests that expect e.g. a certain hash ordering, to use sorted
output or better comparisons.