Don't double-decode geocoded addresses.
Perl 5.20 introduced a version of Encode that errors on decoding already
decoded content (rather than returning the same string). Whilst this can
be taken as a bug in our code (although the decoding exists because some
versions of FastCGI silently UTF-8 encode the content), in the changelog
for Perl the only reference to this change is the line: "Encode has been
upgraded from version 2.49 to 2.60.".