Fix test issue running on PostgreSQL>=9.2 on Mac.
Test::PostgreSQL sets up a new database with initdb and empties the
generated postgresql.conf; PostgreSQL 9.2+ sets its locale/timezone
information at the initdb stage, not when it is run. It is possible
therefore for the tests to be in one timezone but the test database
to be in another (GMT) – causing a test failure when e.g. an update
is created manually. Make sure the first comment is well before the
others being tested, so there is no chance the order gets confused.