Display link to FMS for Councils on reports
When viewing on FMS.com a report made via a council cobrand, this
adds a link to the FMS for Councils page for more info.

This works by breaking out the report meta line ('Reported in the
Street lighting category by Joe Bloggs at 09:51, Tuesday 1 March 2016')
into its own template which can be overridden for the fixmystreet.com cobrand.