Stop using Text::Wrap to wrap email bodies.
If we don't do this, the default quoted-printable encoding will make
sure the lines are physically wrapped anyway. format=flowed could be
an alternative, but I think "=" acting as a quasi-hyphen is somewhat
nicer than every line ending with "=20".

A better solution could be quoted-printable doing soft wrapping only
at word breaks, making each line be readable though end with a " =",
but that didn't seem worth the effort involved.