Reinstate missing margin before reporting form email input
`.form-box--logged-in-name` has no preceding heading, which meant the
grey box background was colliding with whatever content happened to go
before it.

We add some margin, and while we’re at it, remove the extra margin-top
on first-child labels (unnecessary, since .form-boxes have their own
padding) and tidy up the syntax of the .form-box rule set.

Fixes #1418.