Add HTML email templates.
Design is all Zarino. This adds the ability to send HTML emails,
including attached inline images. When included, this is done as a
multipart/related email containing a multipart/alternative (of plain and
HTML) and any attached images, so that the images are available even if
HTML mail is not.

The alert emails list data has been improved so it can be constructed in
the templates rather than the code. Various templates have been tidied.

Various workarounds for email clients have been made, including:
* <th> is used so that the Android 4.x mail client can give them
  `block` styling in the small screen media query.
* Font settings defined on every table cell (<th>) so that sans-serif
  fonts are used in Outlook, rather than Times New Roman.
* A three-column wrapper table to create a 620px centred content area
  that also shrinks down on narrow screens. (Outlook doesn’t like
  max-width, so this is the simplest alternative.)
* Enforcing a sensible (500px) min-width for the main content area,
  on clients that don’t support media queries (eg: native Gmail app).
* Giant borders on buttons so Outlook displays them
* Image alignment with align rather than float.