Feature to create report as body/other user.
This allows users who have the correct permissions to add reports on
behalf of the body or another user.

We enable editing of the email box by default, so that if JavaScript
isn't available, someone can still change the email for the 'another
user' option.

For mysociety/fixmystreetforcouncils#10 and
mysociety/fixmystreetforcouncils#11