Show 'Remove from site' button based on report.
If the user is a member of the body the report was sent to (and the
cobrand has the option enabled), show the remove from site button.
This was previously cobrand-based (it was shown if the user was a
member of the body the cobrand has hardcoded as its council_id),
which was confusing and harder to implement. Fixes #1508.