Improve "Summary" and "Description" form guidance
Form labels now describe what the user should do, and placeholders
give a supporting example of what their input should look like.

Placeholder text in the "Summary" box now suggests that people
include a location in their summary. Likewise with a timescale in
the "Description" box.

Placeholders use potholes as an example, since potholes are the most
common complaint on FMS. Maybe eventually we might want to change the
placeholder text based on the category the user selects…